twitter-client/requirements.txt
grpcio 
googleapis-common-protos
redis
